requirements.txt
google-ads>=28.0.0
protobuf==4.25.1
mcp>=1.2.0
fastmcp
